Refactor import statements to use absolute paths for consistency across modules
