+++ photomise/database/project.py
@@ -1,7 +1,7 @@
 from pathlib import Path
 
 import pendulum
-from database.base import DatabaseManager
+from photomise.database.base import DatabaseManager
 
 from photomise.utilities.logging import setup_logging
