+++ photomise/database/shared.py
@@ -1,4 +1,4 @@
-from database.base import DatabaseManager
+from photomise.database.base import DatabaseManager
 from geopy.distance import great_circle
 
 from photomise.utilities.constants import SHARED_DB_PATH
